Create Auth Module, JWT & Passport
Auth 모듈 생성, JWT 와 Passport 를 활용해 인증 구현, nestJS 의 Guard 미들웨어, Custom Decorator 를 활용하여 User와 Board entity relation 을 구현하였음

+++ src/boards/boards.service.ts
@@ -4,6 +4,7 @@ import { v1 as uuid } from 'uuid';
 import { CreateBoardDto } from './dto/create-board.dto';
 import { Board } from 'src/entities/board.entity';
 import { Repository } from 'typeorm';
+import { User } from 'src/entities/user.entity';
 
 @Injectable()
 export class BoardsService {
@@ -11,15 +12,28 @@ export class BoardsService {
     @Inject('BOARD_REPOSITORY')
     private boardRepository: Repository<Board>,
   ) {}
-
+  d;
   //모든 board 가져오기
   async getAllBoards(): Promise<Board[]> {
-    return await this.boardRepository.find();
+    const boards = await this.boardRepository.find();
+    if (!boards) {
+      throw new NotFoundException('게시글이 존재하지 않습니다.');
+    }
+    return boards;
   }
 
+  // //한 유저의 모든 boards 가져오기
+  // async getAUserBoards(user: User): Promise<Boar[]> {
+  //   const query = this.boardRepository.createQueryBuilder('board');
+  //   query.where('board.userId = :userId', { userId: user.id });
+
+  //   const boards = await query.getMany();
+  //   return boards;
+  // }
+
   //ID 를 통해 board 가져오기
-  async getBoardById(id: number): Promise<Board> {
-    const found = await this.boardRepository.findOneBy({ id });
+  async getBoardById(id: number, user: User): Promise<Board> {
+    const found = await this.boardRepository.findOneBy({ id, userId: user.id });
     if (!found) {
       throw new NotFoundException(`${id} 에 해당하는 것 없음`);
     }
@@ -27,27 +41,35 @@ export class BoardsService {
   }
 
   //board 생성
-  async createBoard(createBoardDto: CreateBoardDto): Promise<Board> {
+  async createBoard(
+    createBoardDto: CreateBoardDto,
+    user: User,
+  ): Promise<Board> {
     const { title, description } = createBoardDto;
 
     const board = this.boardRepository.create({
       title,
       description,
       status: BoardStatus.PUBLIC,
+      user,
     });
 
     return await this.boardRepository.save(board);
   }
 
   //id를 통해 board 삭제하기
-  async deleteBoardById(id: number) {
-    const result = await this.boardRepository.delete(id);
+  async deleteBoardById(id: number, user: User): Promise<void> {
+    const result = await this.boardRepository.delete({ id, userId: user.id });
     console.log('Delete Result : ', result);
-    
   }
+
   //Board 의 공개 상태 업데이트
-  async updateBoardStatus(id: number, status: BoardStatus): Promise<Board> {
-    const board = await this.getBoardById(id);
+  async updateBoardStatus(
+    id: number,
+    status: BoardStatus,
+    user: User,
+  ): Promise<Board> {
+    const board = await this.getBoardById(id, user);
 
     board.status = status;
     await this.boardRepository.save(board);
